:hankey: The serve command now spawns a new process that handles the compiling process this way it can be stopped at any moment
