+++ grafe-cli/src/serve/serve.ts
@@ -3,19 +3,18 @@ import * as chokidar from 'chokidar';
 import * as pkgDir from 'pkg-dir';
 import * as path from 'path';
 import { cat, exec } from 'shelljs';
-import { ChildProcess, ChildProcessWithoutNullStreams, execSync, spawn } from 'child_process';
+import { ChildProcess, ChildProcessByStdio, ChildProcessWithoutNullStreams, execFile, execSync, fork, spawn } from 'child_process';
 import * as fs from 'fs';
 import 'colors';
-import { executionAsyncResource } from 'async_hooks';
-
-let activeProcess: ChildProcess;
-let timeout: NodeJS.Timeout;
-
 
 export function serveCommand(yargs: yargs.Argv<Record<string, unknown>>) {}
 
 export async function serveHandler(argv: Record<string, unknown>) {
     const rootDir = await pkgDir.default(process.cwd());
+
+    let compilerProzess: ChildProcess;
+
+    const compilerPath = path.join(__dirname, 'compiler');
 
     createBuildDirIfNeeded(rootDir);
 
@@ -26,89 +25,19 @@ export async function serveHandler(argv: Record<string, unknown>) {
         .on('all', (event, path) => {
             console.log('\n-- Change detected restarting --\n'.green);
 
-            clearTimeout(timeout);
+            killTask(compilerProzess);
 
-            // maybe it would be better to just fork here but I am not sure
-            // this way if the changes happen within 150ms they will be ignored but there has to be a better way
-            timeout = setTimeout(() => {
-                compile(rootDir);
-            }, 150);
+            compilerProzess = execFile(process.argv[0], [compilerPath, rootDir]);
 
+            compilerProzess.stdout.pipe(process.stdout);
+            compilerProzess.stderr.pipe(process.stderr);
         });
 
-    compile(rootDir);
-}
 
-/**
- * Starts the compiling process. It first deletes everything from the build folder. After that it 
- * starts the Typescript compiler and compiles. After the compiler finished it starts the backend.
- * 
- * This code needs improvemnt and a lot of it...like really a lot
- * @param rootDir The root directory of the project
- */
-async function compile(rootDir: string) {
-    if (activeProcess && !activeProcess.killed) {
-        killTask(activeProcess);
-    }
 
-    try {
-        const deleteDirectory = path.join(rootDir, 'build');
-        console.log('\n-- Cleaning build directory --\n'.green);
-        removeEverythingFromDir(deleteDirectory);
-    } catch (error) {
-        console.error(
-            `\nThere was an error in the cleaning stage: ${error}\n`.red
-        );
-        return;
-    }
-
-    try {
-        console.log(`\n-- Compiling code --\n`.green);
-        const code = await compileCode(rootDir);
-
-        if (code !== 0) {
-            console.error(`\n-- There was an error during compiling. Additional information should be above! Compiler finished with code ${code} --\n`.red);
-            return;
-        }
-
-    } catch (error) {
-        console.error(
-            `\nThere was an error in the compiling stage: ${error}\n`.red
-        );
-        return;
-    }
-
-    try {
-        console.log(`\n-- Starting project --\n`.green);
-        activeProcess = exec('node build/index.js', {
-            cwd: rootDir,
-            async: true,
-        });
-    } catch (error) {
-        console.error(`\nThere was an error in the running stage: ${error}\n`.red);
-        return;
-    }
-}
-
-function removeEverythingFromDir(dir: string) {
-    const files = fs.readdirSync(dir);
-
-    files.forEach((file) => {
-        const filePath = path.join(dir, file);
-        const stats = fs.statSync(filePath);
-
-        if (stats.isFile()) {
-            fs.unlinkSync(filePath);
-        } else if (stats.isDirectory()) {
-            // node 14.x
-            if (typeof fs.rmSync === 'function') {
-                fs.rmSync(filePath, { recursive: true, force: true });
-            } else {
-                // node 12.x
-                fs.rmdirSync(filePath, { recursive: true });
-            }
-        }
-    });
+    compilerProzess = execFile(process.argv[0], [compilerPath, rootDir]);
+    compilerProzess.stdout.pipe(process.stdout);
+    compilerProzess.stderr.pipe(process.stderr);
 }
 
 /**
@@ -132,60 +61,11 @@ function killTask(child: ChildProcess): void {
         exec('taskkill /pid ' + child.pid + ' /T /F', {silent: true});
     } else if (process.platform === 'darwin') {
         // Mac OS
-        child.kill('SIGTERM');
+        child.kill('SIGSTOP');
     } else {
         // Linux
-        child.kill('SIGTERM');
+        child.kill('SIGSTOP');
     }
-}
-
-async function compileCode(rootDir: string) {
-
-    return new Promise<number>((res, rej) => {
-        let compiler: ChildProcessWithoutNullStreams; 
-        
-        compiler = spawn('tsc', {
-            cwd: rootDir,
-            shell: true
-        });
-
-        compiler.on('error', async error => {
-
-            // @ts-ignore
-            if (error.code === 'ENOENT') {
-
-                console.warn('\n-- Unable to find Typescript globally. Falling back to using `npm run build` --'.yellow);
-
-                try {
-                    res(await compileCodeUsingNodeModules(rootDir));
-                } catch(error) {
-                    rej(error);
-                }
-            }
-        });
-        
-        compiler.stdout.on('data', data => {
-            console.log(`${data}`);
-        });
-        
-        compiler.stderr.on('data', data => {
-            console.log(`${data}`);
-        });
-
-        compiler.on('exit', code => {
-            res(code);
-        })
-    });
-}
-
-async function compileCodeUsingNodeModules(rootDir: string): Promise<number> {
-    return new Promise<number>((res, rej) => {
-        const child = exec('npm run build', {
-            cwd: rootDir,
-        });
-
-        res(child.code);
-    });
 }
 
 /**
